requirements.txt
streamlit
pandas
numpy 
matplotlib
dotenv
mysql-connector-python
datetime
hashlib